Optomised the website
I worked on the website to give it the lighthouse treatment, its a chrome dev thing to help make websaites better. I used it to makew the css and js load better and used it to make sure that the seo was good. So that hopefully the page you see when googling the website is not the about us page but the index